- added word lookup by word context (in contrast to lookup by single character occurrence), even on multiple lines - added ability to switch between context lookup and single character lookup by multiple clicks on the same character (context lookup beeing default) - dump log buffer to disk, when a critical exception has been caught, that halts the program

+++ nds/ft_test/source/unicode.h
@@ -7,8 +7,8 @@ class UCChar
 {
 public:
 	unsigned long code_point;
-	unsigned long source_offset;
-	unsigned short source_length;
+	int source_offset;
+	int source_length;
 };
 typedef std::list<UCChar> UCCharList;
